modified:   ParmanodL/ParmanodL_Installer 	modified:   ParmanodL/src/ParmanodL_write.sh

+++ ParmanodL/src/ParmanodL_write.sh
@@ -3,6 +3,13 @@ function ParmanodL_write {
 
 sudo umount -f ${disk}* >/dev/null 2>&1
 
+if [[ $OS == Linux ]] ; then
 sudo dd if="${image_path}" of="${disk}" bs=4194304 status=progress 
+fi
+
+if [[ $OS == Mac ]] ; then
+sudo dd if="${image_path}" of="${disk}" bs=4194304 
+fi
+
 sync
 }
